+++ lib/community/widgets/post_card_metadata.dart
@@ -14,6 +14,7 @@ class PostCardMetaData extends StatelessWidget {
   final int comments;
   final DateTime published;
   final bool saved;
+  final bool distinguised;
 
   const PostCardMetaData({
     super.key,
@@ -22,6 +23,7 @@ class PostCardMetaData extends StatelessWidget {
     required this.comments,
     required this.published,
     required this.saved,
+    required this.distinguised,
   });
 
   final MaterialColor upVoteColor = Colors.orange;
@@ -76,12 +78,12 @@ class PostCardMetaData extends StatelessWidget {
                   text: formatTimeToString(dateTime: published.toIso8601String()),
                 ),
                 const SizedBox(width: 14.0),
-                // if (postView.post.distinguised)
-                // Icon(
-                //   Icons.campaign_rounded,
-                //   size: 24.0,
-                //   color: Colors.green.shade800,
-                // ),
+                if (distinguised)
+                  Icon(
+                    Icons.campaign_rounded,
+                    size: 24.0,
+                    color: Colors.green.shade800,
+                  ),
               ],
             ),
             if (useCompactView)